backend/requirement.txt
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
CouchDB==1.2
debtcollector==1.21.0
Django==2.2
idna==2.8
iso8601==0.1.12
keystoneauth1==3.14.0
msgpack==0.6.1
netaddr==0.7.19
netifaces==0.10.9
os-service-types==1.6.0
oslo.config==6.8.1
oslo.i18n==3.23.1
oslo.serialization==2.28.2
oslo.utils==3.40.3
pbr==5.1.3
Pillow==6.0.0
pyparsing==2.4.0
python-keystoneclient==3.19.0
python-swiftclient==3.7.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
rfc3986==1.3.1
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
ujson==1.35
urllib3==1.24.2
wrapt==1.11.1
